use size_t rather than int for directory lengths in memory this might fix a bug with amd64
svn:r2410

+++ src/or/main.c
@@ -979,7 +979,7 @@ static void dumpstats(int severity) {
 #else
         "Average bandwidth used: %llu/%ld = %d bytes/sec",
 #endif
-        stats_n_bytes_read, stats_n_seconds_uptime,
+        (long long unsigned int)stats_n_bytes_read, stats_n_seconds_uptime,
         (int) (stats_n_bytes_read/stats_n_seconds_uptime));
 
   rep_hist_dump_stats(now,severity);
